+++ src/features/chat/chat-services/chat-thread-service.ts
@@ -257,14 +257,14 @@ export const CreateChatThread = async (): ServerActionResponseAsync<ChatThreadMo
   }
 }
 
-export const InitChatSession = async (
-  props: PromptGPTProps
-): ServerActionResponseAsync<{
+export type InitChatSessionResponse = {
   chatThreadId: string
   updatedLastHumanMessage: ChatMessageModel
   chats: ChatMessageModel[]
   chatThread: ChatThreadModel
-}> => {
+}
+
+export const InitChatSession = async (props: PromptGPTProps): ServerActionResponseAsync<InitChatSessionResponse> => {
   const { messages, id: chatThreadId, chatType, conversationStyle, conversationSensitivity, chatOverFileName } = props
 
   const lastHumanMessage = messages[messages.length - 1]
